refactor: simplificar imports de módulos mediante archivos barril
- Actualizados todos los imports de rutas, controladores y middlewares.
- Eliminadas rutas de archivos específicas en favor de imports centralizados.
- Mejora significativa en la legibilidad y limpieza de las cabeceras de cada archivo.
